refactor: DisciplineCode를 user 패키지에서 event 패키지로 이동 #1
