+++ src/main/java/com/example/trx/domain/event/Dancing.java
@@ -1,7 +1,5 @@
 package com.example.trx.domain.event;
 
-import com.example.trx.domain.user.DisciplineCode;
-
 public class Dancing extends ContestEvent{
 
   public Dancing(Level level) {
