+++ src/main/java/com/example/trx/domain/event/DisciplineCode.java
@@ -1,4 +1,4 @@
-package com.example.trx.domain.user;
+package com.example.trx.domain.event;
 
 public enum DisciplineCode {
     FREESTYLE,
