+++ src/main/java/com/example/trx/domain/event/FreeStyle.java
@@ -1,7 +1,5 @@
 package com.example.trx.domain.event;
 
-import com.example.trx.domain.user.DisciplineCode;
-
 public class FreeStyle extends ContestEvent{
 
   public FreeStyle(Level level) {
