+++ src/main/java/com/example/trx/domain/event/Slalom.java
@@ -1,7 +1,5 @@
 package com.example.trx.domain.event;
 
-import com.example.trx.domain.user.DisciplineCode;
-
 public class Slalom extends ContestEvent{
 
   public Slalom(Level level) {
